@@ -1,40 +1,37 @@
-import * as ESTree from 'estree';
+import { injectable, inject } from 'inversify';
+import { ServiceIdentifiers } from '../container/ServiceIdentifiers';
 
 import { ICustomNode } from '../interfaces/custom-nodes/ICustomNode';
-import { IOptions } from '../interfaces/IOptions';
-import { TStatement } from '../types/TStatement';
+import { IOptions } from '../interfaces/options/IOptions';
+import { TStatement } from '../types/node/TStatement';
 
-import { AppendState } from '../enums/AppendState';
+import { NodeUtils } from '../node/NodeUtils';
 
+@injectable()
 export abstract class AbstractCustomNode implements ICustomNode {
-    /**
-     * @type {AppendState}
-     */
-    protected abstract appendState: AppendState;
-
     /**
      * @type {IOptions}
      */
-    protected options: IOptions;
+    protected readonly options: IOptions;
 
     /**
      * @param options
      */
-    constructor (options: IOptions) {
+    constructor (
+        @inject(ServiceIdentifiers.IOptions) options: IOptions
+    ) {
         this.options = options;
     }
 
     /**
-     * @param astTree
+     * @param args
      */
-    public abstract appendNode (astTree: ESTree.Node): void;
+    public abstract initialize (...args: any[]): void;
 
     /**
-     * @returns {AppendState}
+     * @returns {string}
      */
-    public getAppendState (): AppendState {
-        return this.appendState;
-    }
+    public abstract getCode (): string;
 
     /**
      * @returns {TStatement[]}
@@ -44,14 +41,9 @@ export abstract class AbstractCustomNode implements ICustomNode {
     }
 
     /**
-     * @param appendState
-     */
-    public setAppendState (appendState: AppendState): void {
-        this.appendState = appendState;
-    }
-
-    /**
      * @returns {TStatement[]}
      */
-    protected abstract getNodeStructure (): TStatement[];
+    protected getNodeStructure (): TStatement[] {
+        return NodeUtils.convertCodeToStructure(this.getCode());
+    }
 }
