@@ -1,7 +1,7 @@
 import * as ESTree from 'estree';
 
-import { TNodeWithBlockStatement } from '../types/TNodeWithBlockStatement';
-import { TStatement } from '../types/TStatement';
+import { TNodeWithBlockStatement } from '../types/node/TNodeWithBlockStatement';
+import { TStatement } from '../types/node/TStatement';
 
 import { NodeType } from '../enums/NodeType';
 
