@@ -1,18 +1,21 @@
-import * as chai from 'chai';
+import { ServiceIdentifiers } from '../../../src/container/ServiceIdentifiers';
+
 import * as ESTree from 'estree';
 
-import { TStatement } from '../../../src/types/TStatement';
+import { assert } from 'chai';
 
+import { TStatement } from '../../../src/types/node/TStatement';
+
+import { IInversifyContainerFacade } from '../../../src/interfaces/container/IInversifyContainerFacade';
+import { IStackTraceAnalyzer } from '../../../src/interfaces/stack-trace-analyzer/IStackTraceAnalyzer';
 import { IStackTraceData } from '../../../src/interfaces/stack-trace-analyzer/IStackTraceData';
 
 import { readFileAsString } from '../../helpers/readFileAsString';
 
+import { InversifyContainerFacade } from '../../../src/container/InversifyContainerFacade';
 import { NodeAppender } from '../../../src/node/NodeAppender';
 import { NodeMocks } from '../../mocks/NodeMocks';
 import { NodeUtils } from '../../../src/node/NodeUtils';
-import { StackTraceAnalyzer } from '../../../src/stack-trace-analyzer/StackTraceAnalyzer';
-
-const assert: any = chai.assert;
 
 describe('NodeAppender', () => {
     describe('appendNode (blockScopeNode: TNodeWithBlockStatement[], nodeBodyStatements: TStatement[]): void', () => {
@@ -49,6 +52,10 @@ describe('NodeAppender', () => {
     });
 
     describe('appendNodeToOptimalBlockScope (blockScopeStackTraceData: IStackTraceData[], blockScopeNode: TNodeWithBlockStatement, nodeBodyStatements: TStatement[], index: number = 0): void', () => {
+        const inversifyContainerFacade: IInversifyContainerFacade = new InversifyContainerFacade({});
+        const stackTraceAnalyzer: IStackTraceAnalyzer = inversifyContainerFacade
+            .get<IStackTraceAnalyzer>(ServiceIdentifiers.IStackTraceAnalyzer);
+
         let astTree: ESTree.Program,
             expectedAstTree: ESTree.Program,
             node: TStatement[],
@@ -73,7 +80,7 @@ describe('NodeAppender', () => {
                 )
             );
 
-            stackTraceData = new StackTraceAnalyzer(astTree.body).analyze();
+            stackTraceData = stackTraceAnalyzer.analyze(astTree.body);
             NodeAppender.appendNodeToOptimalBlockScope(stackTraceData, astTree, node);
 
             assert.deepEqual(astTree, expectedAstTree);
@@ -92,7 +99,7 @@ describe('NodeAppender', () => {
                 )
             );
 
-            stackTraceData = new StackTraceAnalyzer(astTree.body).analyze();
+            stackTraceData = stackTraceAnalyzer.analyze(astTree.body);
             NodeAppender.appendNodeToOptimalBlockScope(stackTraceData, astTree, node);
 
             assert.deepEqual(astTree, expectedAstTree);
@@ -118,7 +125,7 @@ describe('NodeAppender', () => {
                     )
                 );
 
-                stackTraceData = new StackTraceAnalyzer(astTree.body).analyze();
+                stackTraceData = stackTraceAnalyzer.analyze(astTree.body);
                 NodeAppender.appendNodeToOptimalBlockScope(stackTraceData, astTree, node, 2);
 
                 assert.deepEqual(astTree, expectedAstTree);
@@ -132,7 +139,7 @@ describe('NodeAppender', () => {
                     )
                 );
 
-                stackTraceData = new StackTraceAnalyzer(astTree.body).analyze();
+                stackTraceData = stackTraceAnalyzer.analyze(astTree.body);
                 NodeAppender.appendNodeToOptimalBlockScope(stackTraceData, astTree, node, 1);
 
                 assert.deepEqual(astTree, expectedAstTree);
@@ -150,7 +157,7 @@ describe('NodeAppender', () => {
                     )
                 );
 
-                stackTraceData = new StackTraceAnalyzer(astTree.body).analyze();
+                stackTraceData = stackTraceAnalyzer.analyze(astTree.body);
                 NodeAppender.appendNodeToOptimalBlockScope(
                     stackTraceData,
                     astTree,
