@@ -2,8 +2,7 @@
 
 var fs = require("fs"),
     nodeExternals = require('webpack-node-externals'),
-    webpack = require('webpack'),
-    ForkCheckerPlugin = require('awesome-typescript-loader').ForkCheckerPlugin;
+    webpack = require('webpack');
 
 function getLicenseText () {
     return "/*\nCopyright (C) 2016 Timofey Kachalov <sanex3339@yandex.ru>\n\n" +
@@ -23,7 +22,6 @@ module.exports = {
                 test: /\.ts(x?)$/,
                 loader: 'awesome-typescript-loader',
                 query: {
-                    forkChecker: true,
                     useBabel: true,
                     useCache: true
                 }
@@ -34,7 +32,6 @@ module.exports = {
         extensions: ['.ts']
     },
     plugins: [
-        new ForkCheckerPlugin(),
         new webpack.BannerPlugin(
             {
                 banner: getLicenseText() + '\n\nrequire("source-map-support").install();\n',
